Split processing helpers into ESPN and video modules
